Terraform Module for ES/KB Helm Chart
A Terraform module is defined under `modules/elastic-cluster/` to invoke
the Helm Chart `charts/elastic-cluster/`, using the TF Helm provider.

Specifically, this module accepts a `values.yaml` file and the deployment
names of Elasticsearch and Kibana to deploy a the defined stack.

Finally, it outputs all Kubernetes internal endpoints of the ES cluster
such as ES service DNS and URL, Kibana DNS and URL, the secret that contains
the `elastic` user credentials, and the TLS certificates