Collector integration with MongoDB in Minikube
The Fluentd collector is used to ship all logs to  the MongoDB
Data Lake deployed in the Minikube example.

For this to happen, a custom docker image was built,
based on `gcr.io/google-containers/fluentd-elasticsearch:v2.4.0`
that contains a build of the native `fluent-plugin-mongo` [1]

The deployment ships all logs to a MongoDB capped collection.