Auditbeat Terraform Module
This commit introduces a minimal Auditbeat Helm Chart
`values.yaml` file, and a Terraform Module that deploys
it through `stable/auditbeat` Helm Chart.

An Auditbeat Daemonset that runs in all K8s nodes
(`Operator: Exists`), is created from the Helm Chart.

The only TF variable is the `collector_url` which is
tailored for use with the `collector` module and its
`collector_beats_url` output.