Elastic-cluster Helm Chart ingresses
This commit introduces Helm templates to create ingresses for
Kibana and Elasticsearch services. Kibana ingress is enabled by default,
while Elasticsearch is not.

The hostnames defined in the chart are `kibana.bluebaron.local`
and `elasticsearch.bluebaron.local`. If not specified otherwise,
the self-signed certificates created with each service are used to enable HTTPS.

Annotations are specified to inform the Ingress controller to forward
ingresses, to HTTPS-enabled backend services.