Rewrite keys on Ingester.
This commit uses fluent-plugin-rename-key [1],
to set rename rules, to revert the sanitization of `$` and `.`
characters in JSON keys, required for creating MongoDB entries.

It does so by replacing `__dot__` and `__dollar__`, delimiters with
the corresponding character.

LIMITATION: The plugin can not create recursive rules,
so only the first character of the string (`$` or `.`)
and the next `.` will be replaced.

[1] https://github.com/shunwen/fluent-plugin-rename-key