Enable  Helm switch for Praeco TF Module
This commit sets `recreate_pods` Helm switch to `true`
for Praeco TF module `modules/praeco/helm.tf`.

This way it is ensured that ConfigMaps and Secrets are remounted
and configuration changes are applied, when re-applying the TF Module.

e.g. Previously when changing the Slack Web-Hook url the baserule_config secret was
recreated, but the elastalert pod kept its previously mounted configuration.