Create an S3 Bucket in the AWS example for on-premise
This commit creates a new Blue Baron example, that starts
the integration of a Blue Baron on-premise deployment with
AWS log sources.

A directory is introduced `examples/aws-on-prem`, where
will reside a deployment of all needed AWS cloud infrastructure
to ingest CloudTrail (and easily extend to other sources),
into a Bare-Metal elasticsearch deployement.