[TF][Praeco] Create variable to accept rule path
As the Praeco `values.yaml` file gets uncontrolably big when 3-4
rules are added in the `elastalert-server.rules` directive,
this change adds the `rules_directory` variable.

This variable, when set to a path, it looks for `.rule.yaml` files
and deploys them as Elastalert rules.

This change is not backwards-compatible, as the rules writen in
`values.yaml` get overriden by the variable.